+++ packages/jest-cli/src/plugins/update_snapshots.js
@@ -31,7 +31,7 @@ class UpdateSnapshotsPlugin extends BaseWatchPlugin {
   }
 
   apply(hooks: JestHookSubscriber) {
-    hooks.testRunComplete(results => {
+    hooks.onTestRunComplete(results => {
       this._hasSnapshotFailure = results.snapshot.failure;
     });
   }
@@ -39,7 +39,7 @@ class UpdateSnapshotsPlugin extends BaseWatchPlugin {
   getUsageInfo(globalConfig: GlobalConfig) {
     if (this._hasSnapshotFailure) {
       return {
-        key: 'u'.codePointAt(0),
+        key: 'u',
         prompt: 'update failing snapshots',
       };
     }