+++ packages/jest-cli/src/types.js
@@ -10,12 +10,8 @@ import type {GlobalConfig} from 'types/Config';
 import type {JestHookSubscriber} from './jest_hooks';
 
 export type UsageData = {
-  key: number,
+  key: string,
   prompt: string,
-};
-
-export type JestHooks = {
-  testRunComplete: any,
 };
 
 export interface WatchPlugin {