+++ packages/jest-diff/src/index.js
@@ -47,8 +47,6 @@ const FALLBACK_FORMAT_OPTIONS_0 = Object.assign({}, FALLBACK_FORMAT_OPTIONS, {
   indent: 0,
 });
 
-const MULTILINE_REGEXP = /[\r\n]/;
-
 // Generate a string that will highlight the difference between two values
 // with green and red. (similar to how github does code diffing)
 function diff(a: any, b: any, options: ?DiffOptions): ?string {
@@ -88,11 +86,7 @@ function diff(a: any, b: any, options: ?DiffOptions): ?string {
 
   switch (aType) {
     case 'string':
-      const multiline = MULTILINE_REGEXP.test(a) && b.indexOf('\n') !== -1;
-      if (multiline) {
-        return diffStrings(a, b, options);
-      }
-      return null;
+      return diffStrings(a, b, options);
     case 'number':
     case 'boolean':
       return null;