+++ packages/jest-editor-support/index.d.ts
@@ -179,6 +179,6 @@ export interface SnapshotMetadata {
 }
 
 export class Snapshot {
-  constructor(parser: any, customMatchers?: string[]);
+  constructor(parser?: any, customMatchers?: string[]);
   getMetadata(filepath: string): SnapshotMetadata[];
 }