+++ packages/jest-environment-jsdom/src/__mocks__/index.js
@@ -14,7 +14,7 @@ JSDOMEnvironment.mockImplementation(function(config) {
   this.global = {
     JSON,
     console: {},
-    mockClearTimers: jest.genMockFn(),
+    mockClearTimers: jest.fn(),
   };
 
   const globalValues = Object.assign({}, config.globals);