+++ packages/jest-jasmine2/src/expectation_result_factory.js
@@ -36,12 +36,18 @@ function stackFormatter(options, initError, errorMessage) {
     return '';
   }
 
-  if (options.error && options.error.stack) {
-    return options.error.stack;
+  if (options.error) {
+    if (options.error.stack) {
+      return options.error.stack;
+    }
+
+    if (options.error === errorMessage) {
+      return errorMessage;
+    }
   }
 
   if (initError) {
-    return errorMessage + '\n' + initError.stack;
+    return errorMessage.trimRight() + '\n\n' + initError.stack;
   }
 
   return new Error(errorMessage).stack;