+++ packages/jest-jasmine2/src/index.js
@@ -15,7 +15,6 @@ import type {TestResult} from 'types/TestResult';
 import type Runtime from 'jest-runtime';
 
 import path from 'path';
-import fs from 'graceful-fs';
 import installEach from './each';
 import {getCallsite} from 'jest-util';
 import JasmineReporter from './reporter';
@@ -119,30 +118,6 @@ async function jasmine2(
     runtime.requireModule(config.setupTestFrameworkScriptFile);
   }
 
-  runtime
-    .requireInternalModule(
-      require.resolve('source-map-support'),
-      'source-map-support',
-    )
-    .install({
-      environment: 'node',
-      handleUncaughtExceptions: false,
-      retrieveSourceMap: source => {
-        const sourceMaps = runtime.getSourceMaps();
-        const sourceMapSource = sourceMaps && sourceMaps[source];
-
-        if (sourceMapSource) {
-          try {
-            return {
-              map: JSON.parse(fs.readFileSync(sourceMapSource)),
-              url: source,
-            };
-          } catch (e) {}
-        }
-        return null;
-      },
-    });
-
   if (globalConfig.enabledTestsMap) {
     env.specFilter = spec => {
       const suiteMap =
@@ -157,9 +132,10 @@ async function jasmine2(
 
   runtime.requireModule(testPath);
   await env.execute();
-  return reporter
-    .getResults()
-    .then(results => addSnapshotData(results, snapshotState));
+
+  const results = await reporter.getResults();
+
+  return addSnapshotData(results, snapshotState);
 }
 
 const addSnapshotData = (results, snapshotState) => {